Missing function `toggle-read-only`
There seems to be a missing function called `toggle-read-only` that's used in
several places throughout the codebase. Not sure if this was once support in a
previous version of emacs or a dependent package, but in the version that I am
using (29.3) it is not available.

I've added a simple implementation to make the code work.